requirements.txt
psycopg2
numpy
flask
python-dotenv

